FIX: API json decode error
