+++ ExpenseTracker/Preview Content/PreviewData.swift
@@ -8,6 +8,9 @@
 import Foundation
 import SwiftUI
 
-var transactionPreviewData = Transaction(id: 1, date: "01/24/2022", institution: "Desjardins", account: "Visa Desjardins", merchant: "Apple", amount: 11.49, type: "debit", categoryId: 801, category: "Software", isPanding: false, isTransfer: false, isExpense: true, isEdited: false)
+//var transactionPreviewData = Transaction(id: 1, date: "01/24/2022", institution: "Desjardins", account: "Visa Desjardins", merchant: "Apple", amount: 11.49, type: "debit", categoryId: 801, category: "Software", isPanding: false, isTransfer: false, isExpense: true, isEdited: false)
 
-var transationListPreviewData = [Transaction](repeating: transactionPreviewData, count: 10)
+
+//var transactionPreviewData = Transaction(id: 1, date: "01/24/2022", institution: "Desjardins", account: Account(rawValue: "Visa Desjardins") ?? <#default value#>, merchant: "Apple", amount: 11.49, type: "debit", categoryID: 801, category: "Software", isPending: false, icon: false)
+//
+//var transationListPreviewData = [Transaction](repeating: transactionPreviewData, count: 10)
